+++ src/game/dialog/FacialExpressions.java
@@ -1,35 +1,35 @@
 package game.dialog;
 
 /**
- * This class holds the integer values of the facial expressions used by the
- * {@link game.dialog.Dialog} class.
+ * This class holds the animation identification values of the facial
+ * expressions used by the {@link game.dialog.Dialog} class.
  * 
  * @author Albert Beaupre
  */
 public interface FacialExpressions {
 
-	int REALLY_SAD = 9760;
-	int SAD = 9765;
-	int DEPRESSED = 9770;
-	int WORRIED = 9775;
-	int SCARED = 9780;
-	int MEAN_FACE = 9785;
-	int MEAN_HEAD_BANG = 9790;
-	int EVIL = 9795;
-	int FREAKING_OUT = 9800;
-	int NON_TALKING = 9805;
-	int CALM_TALK = 9810;
-	int TOUGH = 9815;
-	int SNOBBY = 9820;
-	int SNOBBY_HEAD_MOVE = 9825;
-	int CONFUSED = 9830;
-	int DRUNK_HAPPY_T = 9835;
-	int TALKING_FAST = 9845;
-	int HAPPY = 9850;
-	int BAD_ASS = 9855;
-	int THINKING = 9860;
-	int COOL_YES = 9864;
-	int EXCITEDLY_LAUGHING = 9851;
-	int TALKING_SECRETLY = 9838;
+	short REALLY_SAD = 9760;
+	short SAD = 9765;
+	short DEPRESSED = 9770;
+	short WORRIED = 9775;
+	short SCARED = 9780;
+	short MEAN_FACE = 9785;
+	short MEAN_HEAD_BANG = 9790;
+	short EVIL = 9795;
+	short FREAKING_OUT = 9800;
+	short NON_TALKING = 9805;
+	short CALM_TALK = 9810;
+	short TOUGH = 9815;
+	short SNOBBY = 9820;
+	short SNOBBY_HEAD_MOVE = 9825;
+	short CONFUSED = 9830;
+	short DRUNK_HAPPY_T = 9835;
+	short TALKING_FAST = 9845;
+	short HAPPY = 9850;
+	short BAD_ASS = 9855;
+	short THINKING = 9860;
+	short COOL_YES = 9864;
+	short EXCITEDLY_LAUGHING = 9851;
+	short TALKING_SECRETLY = 9838;
 
 }
